+++ changelog.md
@@ -4,6 +4,7 @@ Upcoming (TBD)
 Features
 --------
 * Preserve the query as metadata when saving to Parquet with `.>`.
+* Make completion candidate match order configurable.
 
 
 2.23.0 (2026/09/09)